This pull request introduces significant refactoring to the game logic
and UI, focusing on removing the consultant and call-for-tenders
concepts and replacing them with projects and formation programs. It
also updates salary calculation for employees, adjusts game creation
parameters, and streamlines the build and release workflow. The most
important changes are grouped below by theme.

**Game Logic and Data Model Updates:**

* Consultants and call-for-tenders have been removed from both the UI
(`CurrentGameScreen.cs`) and the database schema, replaced by projects
and formation programs. A new migration (`RemoveConsultants.cs`) drops
consultant-related columns and updates table naming conventions.
[[1]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL414-R415)
[[2]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL490-R491)
[[3]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL619-R630)
[[4]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL719-R756)
[[5]](diffhunk://#diff-765a441147392d7aa12234ae4760479ad976329a83923361f01290a9ab5a3c23R1-R104)
* Employee salary is now calculated dynamically based on skill levels
rather than a fixed random range, and the salary type is changed from
`int` to `double` for more precision.
[[1]](diffhunk://#diff-93e33e87b043d4f7c79812457bb4888cf34da86f3c0e102abde034a93064405fL71-R94)
[[2]](diffhunk://#diff-8829b812785c8721d85fb3e711833380de722e2a3db69f3b0950df9d18c4b472L39-R39)

**Game Actions and UI Refactor:**

* Game actions have been updated to reflect the new concepts: actions
for participating in projects and enrolling in formation replace
previous consultant and call-for-tenders actions, with corresponding UI
text changes. The action for ending a round is now called "Confirm My
Turn."
[[1]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL736-R783)
[[2]](diffhunk://#diff-36d32944f51c67cb0662ded13f7fe044b63e7b8dc6b82b4eab6556e776c7a99fL781-R827)
[[3]](diffhunk://#diff-c305a7ee3a858f71d851d7c5d6f91854c0c05c48c5e4af2f236dcff5d71bc249L50-R50)

**Game Creation Parameters:**

* The default number of rounds for a new game is now 5 (was 15), and
validation allows between 5 and 20 rounds.
[[1]](diffhunk://#diff-4920131ab7e5908730c28ae2bfd96a5e7827293618c1fc30112f0465ed1ad0f4L12-R12)
[[2]](diffhunk://#diff-4920131ab7e5908730c28ae2bfd96a5e7827293618c1fc30112f0465ed1ad0f4L21-R21)

**Build and Release Workflow Improvements:**

* The GitHub workflow for changelog and release has been refactored:
.NET version bumped to 9.x, EF tools installation added, multi-platform
build steps consolidated, and semantic-release is now run after artifact
packaging. Workflow permissions are also updated to include issues.
[[1]](diffhunk://#diff-a0aba5a2bd30b2f80a3a18aa24ce43670b1320ca0a8b9c103bb3bc971ae30a85R12)
[[2]](diffhunk://#diff-a0aba5a2bd30b2f80a3a18aa24ce43670b1320ca0a8b9c103bb3bc971ae30a85L32-R78)

**Miscellaneous:**

* Debug logging added to server round action application for easier
tracing.

+++ Server/Models/Employee.cs
@@ -2,13 +2,13 @@ using Server.Hubs.Records;
 
 namespace Server.Models;
 
-public class Employee(string name, int companyId, int gameId, int salary)
+public class Employee(string name, int companyId, int gameId, double salary)
 {
     public int? Id { get; private set; }
 
     public string Name { get; set; } = name;
 
-    public int Salary { get; set; } = salary;
+    public double Salary { get; set; } = salary;
 
     public ICollection<LeveledSkill> Skills { get; } = [];
